Fix the order of the database cleanup code

+++ playwright/fixtures/supabase-fixture.ts
@@ -46,37 +46,40 @@ export const test = base.extend<{
 			if (createdUsers.length > 0) {
 				const supabaseAdmin = getSupabaseAdmin();
 
-				for (const user of createdUsers) {
+				for await (const user of createdUsers) {
 					try {
-						// delete decks
-						const deletedDecks = await supabaseAdmin
+						// the data must be deleted in a certain order to avoid "foreign key constraint" errors
+
+						// delete deck attribute types
+						const decks = await supabaseAdmin
 							.from("deck")
-							.delete()
-							.eq("user_id", user.id)
-							.select();
-						console.log("deleted decks", deletedDecks);
-						deletedDecks.data?.forEach(async (deck) => {
-							// delete deck attribute types
+							.select()
+							.eq("user_id", user.id);
+						for await (const deck of decks.data ?? []) {
 							await supabaseAdmin
 								.from("deck_attribute_type")
 								.delete()
 								.eq("deck_id", deck.id);
 
-							// delete cards
-							const deletedCards = await supabaseAdmin
+							// delete card attribute values
+							const cards = await supabaseAdmin
 								.from("card")
-								.delete()
-								.eq("deck_id", deck.id)
-								.select();
-
-							// delete card attributes values
-							deletedCards.data?.forEach(async (card) => {
+								.select()
+								.eq("deck_id", deck.id);
+							cards.data?.forEach(async (card) => {
 								await supabaseAdmin
 									.from("card_attribute_value")
 									.delete()
 									.eq("card_id", card.id);
 							});
-						});
+
+							// delete cards
+							await supabaseAdmin.from("card").delete().eq("deck_id", deck.id);
+						}
+
+						// delete decks
+						await supabaseAdmin.from("deck").delete().eq("user_id", user.id);
+
 						// delete user
 						await supabaseAdmin.auth.admin.deleteUser(user.id);
 					} catch (error) {
